backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python
tensorflow
torch
torchvision
matplotlib
ultralytics
pillow
python-dotenv